feat: add toilet dumping support
