+++ bukkit/build.gradle
@@ -9,6 +9,7 @@ dependencies {
     implementation 'io.papermc:paperlib:1.0.8'
     implementation 'space.arim.morepaperlib:morepaperlib:0.4.4'
     implementation 'net.william278.cloplib:cloplib-bukkit:2.0.3'
+    implementation 'net.william278.toilet:toilet-bukkit:1.0.1'
     implementation 'org.bstats:bstats-bukkit:3.1.0'
     implementation 'de.exlll:configlib-yaml:4.5.0'
