+++ common/build.gradle
@@ -25,6 +25,7 @@ dependencies {
     compileOnly 'com.google.guava:guava:33.4.0-jre'
     compileOnly 'org.projectlombok:lombok:1.18.36'
 
+    compileOnlyApi 'net.william278.toilet:toilet-common:1.0.1'
     compileOnlyApi 'net.william278.huskhomes:huskhomes-common:4.9.2-473c5e7'
 
     compileOnly 'net.luckperms:api:5.4'
