+++ fabric/build.gradle
@@ -17,6 +17,7 @@ dependencies {
     modImplementation include("eu.pb4:placeholder-api:${fabric_placeholder_api_version}")
     modImplementation include("net.impactdev.impactor.api:economy:${fabric_impactor_api_version}")
     modImplementation include("net.william278.cloplib:cloplib-fabric:2.0.3+$project.name")
+    modImplementation include("net.william278.toilet:toilet-fabric:1.0.1+$project.name")
     modImplementation "net.fabricmc.fabric-api:fabric-api:${fabric_api_version}"
 
     implementation include('org.apache.commons:commons-pool2:2.12.1')
